Added code for Mongo Controller

+++ src/main/java/com/spring/mongodb/models/User.java
@@ -12,6 +12,6 @@ import org.springframework.data.mongodb.core.mapping.Document;
 public class User {
     private String name;
     private int age;
-    private int bmi;
+    private double bmi;
     private Address address;
 }
